Conflicts:
	tajo-core/src/main/java/org/apache/tajo/querymaster/Query.java
	tajo-core/src/main/java/org/apache/tajo/util/IndexUtil.java

+++ tajo-core/src/main/java/org/apache/tajo/engine/utils/TupleCacheScanner.java
@@ -99,7 +99,7 @@ public class TupleCacheScanner implements Scanner {
     if (cacheData.size() == 0) {
       return 1.0f;
     }
-    return count / cacheData.size();
+    return ((float)count) / cacheData.size();
   }
 
   @Override
